% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{build_loo_metaweb}
\alias{build_loo_metaweb}
\title{Metaweb built from all networks except one focal network}
\usage{
build_loo_metaweb(network_list, exclude_index)
}
\arguments{
\item{network_list}{A named list of matrices, ordered by time point.}

\item{exclude_index}{Index of the network to leave out.}
}
\value{
A matrix: the union metaweb of every network except
\code{network_list[[exclude_index]]}.
}
\description{
Metaweb built from all networks except one focal network
}
