% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{calc_contribution}
\alias{calc_contribution}
\title{Species and interaction overlap of a focal network with a reference metaweb}
\usage{
calc_contribution(focal_net, reference_meta)
}
\arguments{
\item{focal_net}{A bipartite matrix.}

\item{reference_meta}{A reference metaweb matrix, e.g. from
\code{\link[=build_metaweb]{build_metaweb()}} or \code{\link[=build_loo_metaweb]{build_loo_metaweb()}}.}
}
\value{
A list with species/interaction counts and their overlap with
\code{reference_meta}.
}
\description{
Species and interaction overlap of a focal network with a reference metaweb
}
