% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_richness}
\alias{plot_richness}
\title{Species richness of both trophic levels over time}
\usage{
plot_richness(results)
}
\arguments{
\item{results}{A \code{bd_pipeline_results} from \code{\link[=run_network_pipeline]{run_network_pipeline()}}.}
}
\value{
A ggplot object.
}
\description{
Species richness of both trophic levels over time
}
