% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_turnover}
\alias{plot_turnover}
\title{Turnover between consecutive periods: Jaccard + gains/losses}
\usage{
plot_turnover(results)
}
\arguments{
\item{results}{A \code{bd_pipeline_results} from \code{\link[=run_network_pipeline]{run_network_pipeline()}}.}
}
\value{
A list of two ggplot objects (Jaccard similarity; gains/losses).
}
\description{
Turnover between consecutive periods: Jaccard + gains/losses
}
