% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.bd_pipeline_results}
\alias{print.bd_pipeline_results}
\title{Print a bd_pipeline_results object}
\usage{
\method{print}{bd_pipeline_results}(x, ...)
}
\arguments{
\item{x}{A \code{bd_pipeline_results} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a bd_pipeline_results object
}
